+++ apps/ui/src/app/layout.tsx
@@ -3,6 +3,7 @@ import { Geist, Geist_Mono, Press_Start_2P } from "next/font/google";
 import "./globals.css";
 import { SolanaProvider } from "@/providers/SolanaProvider";
 import { AudioProvider } from "@/providers/AudioProvider";
+import { PrivyProvider } from "@/providers/PrivyProvider";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -96,9 +97,11 @@ export default function RootLayout({
       <body
         className={`${geistSans.variable} ${geistMono.variable} ${pressStart.variable}`}
       >
-        <SolanaProvider>
-          <AudioProvider>{children}</AudioProvider>
-        </SolanaProvider>
+        <PrivyProvider>
+          <SolanaProvider>
+            <AudioProvider>{children}</AudioProvider>
+          </SolanaProvider>
+        </PrivyProvider>
       </body>
     </html>
   );